Fix lecture examples and exercise alignment

+++ week01/lectures/lecture_exercises/week01_starter.c
@@ -1,12 +1,13 @@
+#include <limits.h>
 #include <stdio.h>
 
 int main(void) {
   int value;
-  int sum = 0;
+  long long sum = 0;
 
   /* TODO: read integers until input ends and update sum when appropriate. */
   (void)value;
 
-  printf("%d\n", sum);
+  printf("%lld\n", sum);
   return 0;
 }
